requirements.txt
asgiref==3.3.1
astroid==2.4.2
cffi==1.14.4
colorama==0.4.4
cryptography==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.5
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
pytz==2020.5
PyJWT
setuptools
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1